requirements.txt
Django
djangorestframework
psycopg2
python-dotenv
gunicorn
whitenoise
channels
channels-redis
channels_redis